separate code generation from rest of function implementation
