+++ src/api.cpp
@@ -8,7 +8,7 @@
 #include <iostream>
 #define PRINT_LEXEMES
 #define PRINT_AST
-//#define PRINT_CODE
+#define PRINT_CODE
 #endif
 
 using jarl::vm;
@@ -49,15 +49,13 @@ void jarl::execute(vm v, const char* code){
     for(auto& error: errors){
       v->print(error.get());
     }
-    //return;
+    return;
   }
   
   #ifdef PRINT_AST
   std::cout << "::AST::" << std::endl;
   std::cout << parse_tree->toStrDebug() << std::endl;
   #endif
-  
-  return;
   
   Function* proc = new Function(parse_tree, &errors);
   if(errors.size() > 0){
@@ -72,6 +70,8 @@ void jarl::execute(vm v, const char* code){
   std::cout << "::proc::\n";
   std::cout << proc->opcodesToStrDebug() << std::endl;
   #endif
+  
+  return;
   
   v->execute(*proc);
 }
